# specs/api.yaml
openapi: 3.0.0
info:
  title: Task Management API
  description: API para un sistema de gestión de tareas.
  version: 1.0.0
paths:
  /usuarios:
    get:
      summary: Listar todos los usuarios
      description: Requiere privilegios de 'admin'.
      operationId: getUsers
      responses:
        "200":
          description: Una lista de usuarios.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
        "401":
          description: Acceso no autorizado (token inválido o ausente).
        "403":
          description: Acceso denegado (el usuario no tiene rol admin).
  /usuarios/{id}:
    delete:
      summary: Borrar un usuario existente
      description: Requiere privilegios de 'admin'.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: El identificador único del usuario
      responses:
        "204":
          description: Usuario borrado exitosamente.
        "401":
          description: Acceso no autorizado (token inválido o ausente).
        "403":
          description: Acceso denegado (el usuario no tiene rol admin).
        "404":
          description: Usuario no encontrado.
  /tareas:
    get:
      summary: Listar todas las tareas
      operationId: getTasks
      responses:
        "200":
          description: Una lista de tareas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      summary: Crear una tarea nueva
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Tarea creada exitosamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Datos de entrada inválidos.
        "409":
          description: Conflicto. Ya existe una tarea con ese título.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /tareas/{id}:
    get:
      summary: Obtener los detalles de una tarea específica
      operationId: getTaskById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: El identificador único de la tarea
      responses:
        "200":
          description: Detalles de la tarea.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Tarea no encontrada.
    put:
      summary: Actualizar detalles de una tarea existente
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: El identificador único de la tarea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Tarea actualizada exitosamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Datos de entrada inválidos.
        "404":
          description: Tarea no encontrada.
  /registro:
    post:
      summary: Registrar nuevo usuario
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: Usuario registrado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Faltan datos o usuario ya existe.
  /login:
    post:
      summary: Iniciar sesión
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "401":
          description: Credenciales inválidas
  /dashboard:
    get:
      summary: Obtener estadísticas globales de todos los usuarios
      operationId: getGlobalDashboardStats
      responses:
        "200":
          description: Estadísticas totales del sistema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estadisticas"
  /dashboard/personal:
    get:
      summary: Obtener estadísticas personales del usuario autenticado
      operationId: getPersonalDashboardStats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Estadísticas personales del usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estadisticas"
        "401":
          description: No autorizado
  /usuarios/{id}/tareas:
    get:
      summary: Obtener tareas asignadas a un usuario específico
      operationId: getUserTasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "404":
          description: Usuario no encontrado
components:
  schemas:
    Usuario:
      type: object
      required: [id, username, password_hash]
      properties:
        id:
          type: string
        username:
          type: string
        password_hash:
          type: string
        rol:
          type: string
          enum: [admin, user]
          default: user
    Estadisticas:
      type: object
      required: [tareas_completadas, tareas_pendientes, tasa_exito]
      properties:
        tareas_completadas:
          type: number
        tareas_pendientes:
          type: number
        tasa_exito:
          type: string
    Task:
      type: object
      required:
        - id
        - titulo
        - estado
      properties:
        id:
          type: string
          description: Identificador único de la tarea.
        titulo:
          type: string
          description: Título o descripción corta de la tarea.
        estado:
          type: string
          description: Estado actual de la tarea.
          enum: [pendiente, en_progreso, completada]
        descripcion:
          type: string
          description: Descripción extendida del objetivo o tarea.
        horas:
          type: number
          description: Horas estimadas o dedicadas.
        ubicacion:
          type: string
          description: Ubicación o enlace relacionado con la tarea.
        usuario_id:
          type: string
          description: ID del usuario creador y propietario de esta tarea.
    TaskInput:
      type: object
      required:
        - titulo
      properties:
        titulo:
          type: string
          description: Título o descripción corta de la tarea.
        estado:
          type: string
          description: Estado inicial de la tarea. Por defecto será 'pendiente'.
          enum: [pendiente, en_progreso, completada]
          default: pendiente
        descripcion:
          type: string
        horas:
          type: number
        ubicacion:
          type: string
        usuario_id:
          type: string
          description: ID del usuario al que se asignará.
    TaskUpdate:
      type: object
      properties:
        titulo:
          type: string
        estado:
          type: string
          enum: [pendiente, en_progreso, completada]
        descripcion:
          type: string
        horas:
          type: number
        ubicacion:
          type: string